handlers done?

+++ cmd/gophermart/storage/db/migrations/00002_users_orders_table.sql
@@ -1,9 +1,9 @@
 -- +goose Up
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS users_orders (
-    id     SERIAL PRIMARY KEY,
-    login  TEXT   NOT NULL,
-    orders INTEGER[] NOT NULL,
+    id      SERIAL PRIMARY KEY,
+    login   TEXT   NOT NULL,
+    orders  BIGINT[] NOT NULL,
     FOREIGN KEY (login) REFERENCES users(login)
 );
 -- +goose StatementEnd
